Fixed #include names, changed namespace random to kingdom_random
I don't know how code::blocks does includes, but a lot of the names had the wrong case, or were actually in a different directory.

Having the random namespace overwrote libstdc++'s (not sure about clang's libc++) function random, causing a type error (can't overwrite a function with a namespace), so I changed all instances of random to kingdom_random so that there wouldn't be a conflict.